chore: correct config mishaps
